Order history (100%) Unregistered user checkout(80%)
Missing email with key(s) and product(s) after purchase

+++ WebApplication1/WebApplication1/Pages/AddKey.cshtml.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using WebApplication1.Data;
+using WebApplication1.Resource;
 using static Microsoft.EntityFrameworkCore.DbLoggerCategory;
 
 namespace WebApplication1.Pages
@@ -42,35 +43,60 @@ namespace WebApplication1.Pages
 
             var id = _userManager.GetUserId(User);
             mehdiid = id;
-            var fullcart = (from cartprd in _context.Shopping_Card_Products
-                           from cartid in _context.Shopping_card
-                           where cartprd.Shopping_card_ID == cartid.ID && cartid.User_ID == id
-                           select cartprd);
+            if (User.Identity.IsAuthenticated)
+            {
+                var fullcart = (from cartprd in _context.Shopping_Card_Products
+                                from cartid in _context.Shopping_card
+                                where cartprd.Shopping_card_ID == cartid.ID && cartid.User_ID == id
+                                select cartprd);
+                proptabtab = fullcart.ToList();
 
-            proptabtab = fullcart.ToList();
+                foreach (var item in proptabtab)
+                {
+                    while (i < item.quantity)
+                    {
+                        Key keyz = new Key()
+                        {
+                            UserID = id,
+                            License = Guid.NewGuid().ToString(),
+                            ProductID = item.Product_ID
+                        };
+                        _context.Key.Add(keyz);
+                        i = i + 1;
+                    }
+                    i = -1;
+                }
+
+                _context.Shopping_Card_Products.RemoveRange(fullcart);
+
+            }
+
+            else
+            {
+                string cookieshoping = Request.Cookies["ShoppingCart"];
+                List<shoppingCart_cookie> shoppingcartlist = Cookie.Cookiereader_shoppingcart(cookieshoping);
+
+                foreach (shoppingCart_cookie itemm in shoppingcartlist)
+                {
+                    while (i < itemm.Quantity - 1)
+                    {
+                        Key keyz = new Key()
+                        {
+                            //UserID = id,
+                            License = Guid.NewGuid().ToString(),
+                            ProductID = itemm.ProductID
+                        };
+                        _context.Key.Add(keyz);
+                        i = i + 1;
+                    }
+                    i = -1;
+                }
+            }
+
             if (!ModelState.IsValid)
             {
                 return Page();
             }
-
-            
-            foreach (var item in proptabtab)
-            {
-                while (i < item.quantity)
-                {
-                    Key keyz = new Key()
-                    {
-                        UserID = id,
-                        License = Guid.NewGuid().ToString(),
-                        ProductID = item.Product_ID
-                    };
-                    _context.Key.Add(keyz);
-                    i = i + 1;
-                }
-                i = -1;
-            }
-
-            _context.Shopping_Card_Products.RemoveRange(fullcart);
 
             await _context.SaveChangesAsync();
             return RedirectToPage("./Index");
